Wire contracts/shared/rate_curve.rs into the shared crate
Closes #1109
