+++ contracts/shared/src/lib.rs
@@ -1,2 +1,3 @@
 #![no_std]
+extern crate alloc;
 pub mod assets; pub mod auth; pub mod errors; pub mod events; pub mod rate_curve; pub mod sanitizer; pub mod types; pub mod validation;
