+++ contracts/shared/src/sanitizer.rs
@@ -1,3 +1,5 @@
+use alloc::string::String;
+
 /// Removes null bytes and control characters from a memo.
 pub fn sanitize_memo(input: String) -> String {
     input
